### 2023-04-07 #5347 @danielskatz
@timbode - thanks for your submission.

You need to add a line with `---` after the header, before the paper body, as in the example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Please do this, and then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

Also, due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.
