### 2023-04-07 #5347 @danielskatz
It would also be helpful if you added some text to the paper to explain what the software is used for, remembering that JOSS editors, reviewers, and readers come from very diverse fields.
