### 2023-04-11 #5347 @timbode
> You need to add a line with `---` after the header, before the paper body, as in the example paper: https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography

Done, thanks for pointing this out.

> Please do this, and then use the command `@editorialbot generate pdf` to make a new PDF. editorialbot commands need to be the first entry in a new comment.

I have posted the command.

> Also, due to the relatively small amount of code, the editors will now discuss if it meets the [substantial scholarly effort](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort) criterion for review by JOSS. You should hear back in a week or so.

I'd like to point out that the scholarly effort is not entirely reflected in the git history, as large prototypical parts of the code have been developed by us elsewhere. Also, the main contribution is the algorithmic implementation per se, while the conciseness of the code is rather a feature, from our perspective.
