### 2023-04-13 #5364 @editorialbot
[REVIEW]: QAOA.jl: Toolkit for the Quantum and Mean-Field Approximate Optimization Algorithms
**Submitting author:** @timbode (<a href="http://orcid.org/0000-0001-8280-3891">Tim Bode</a>)
**Repository:** https://github.com/FZJ-PGI-12/QAOA.jl
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.1
