### 2023-04-13 #5364 @danielskatz
👋 @timbode - Please fix the missing DOI that editorialbot suggests. Please feel free to make changes to your .bib file, then use the command `@editorialbot check references` to check again, and the command `@editorialbot generate pdf` when the references are right to make a new PDF. editorialbot commands need to be the first entry in a new comment.
