### 2023-04-15 #5364 @Abinashbunty
> * [ ]  **Contribution and authorship:** Has the submitting author (@timbode) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

For the referenced checklist, I am going to mark it as ☑️ based on https://github.com/openjournals/joss-reviews/issues/5347#issuecomment-1502813847 and the reaction to it by @danielskatz
