### 2023-04-18 FZJ-PGI-12/QAOA.jl#26 @Abinashbunty
Missing DOI suggestion from editorial bot for JOSS
Hi author(s).
For `Yao.jl`, the suggestion from editorial bot can be checked. 😄 For reference, here is the comment: https://github.com/openjournals/joss-reviews/issues/5364#issuecomment-1507134220
