### 2023-04-19 9fa1eaa
Fix: Typo in `Summary` section (#25)
