+++ paper/paper.md
@@ -33,7 +33,7 @@ bibliography: paper.bib
 
 Quantum algorithms are an area of intensive research thanks to their potential of speeding up certain specific tasks exponentially. However, for the time being, high error rates on the existing hardware realizations preclude the application of many algorithms that are based on the assumption of fault-tolerant quantum computation. On such _noisy intermediate-scale quantum_ (NISQ) devices [@Preskill2018], the exploration of the potential of _heuristic_ quantum algorithms has attracted much interest. A leading candidate for solving combinatorial optimization problems is the so-called _Quantum Approximate Optimization Algorithm_ (QAOA) [@Farhi:2014].
 
-`QAOA.jl` is a `Julia` package [@bezanson2017julia] package that implements the _mean-field Approximate Optimization Algorithm_ (mean-field AOA) [@mf_aoa] - a quantum-inspired classical algorithm derived from the QAOA via the mean-field approximation. This novel algorithm is useful in assisting the search for quantum advantage by providing a tool to discriminate (combinatorial) optimization problems that can be solved classically from those that cannot.  Note that `QAOA.jl` has already been used during the research leading to [@mf_aoa].
+`QAOA.jl` is a `Julia` package [@bezanson2017julia] that implements the _mean-field Approximate Optimization Algorithm_ (mean-field AOA) [@mf_aoa] - a quantum-inspired classical algorithm derived from the QAOA via the mean-field approximation. This novel algorithm is useful in assisting the search for quantum advantage by providing a tool to discriminate (combinatorial) optimization problems that can be solved classically from those that cannot.  Note that `QAOA.jl` has already been used during the research leading to [@mf_aoa].
 
 Additionally, `QAOA.jl`  also implements the QAOA efficiently to support the extensive classical simulations typically required in research on the topic. The corresponding parameterized circuits are based on `Yao.jl` [@YaoFramework2019], [@Yao] and `Zygote.jl` [@ZygoteFramework], [@Zygote], making it both fast and automatically differentiable, thus enabling gradient-based optimization. A number of common optimization problems such as MaxCut, the minimum vertex-cover problem, the Sherrington-Kirkpatrick model, and the partition problem are pre-implemented to facilitate scientific benchmarking.
 
