+++ paper/paper.bib
@@ -44,11 +44,16 @@
 }
 
 @article{YaoFramework2019,
-  Title={Yao.jl: Extensible, Efficient Framework for Quantum Algorithm Design},
-  Author={Xiu-Zhe Luo and Jin-Guo Liu and Pan Zhang and Lei Wang},
-  Eprint = {1912.10877},
-  Journal = {ArXiv e-prints},
-  Year={2019}
+  doi = {10.22331/q-2020-10-11-341},
+  url = {https://doi.org/10.22331/q-2020-10-11-341},
+  title = {Yao.jl: {E}xtensible, {E}fficient {F}ramework for {Q}uantum {A}lgorithm {D}esign},
+  author = {Luo, Xiu-Zhe and Liu, Jin-Guo and Zhang, Pan and Wang, Lei},
+  journal = {{Quantum}},
+  publisher = {{Verein zur F{\"{o}}rderung des Open Access Publizierens in den Quantenwissenschaften}},
+  volume = {4},
+  pages = {341},
+  month = oct,
+  year = {2020}
 }
 
 @misc{Yao,
