### 2023-04-19 #5364 @Abinashbunty
> * [x]  **Substantial scholarly effort:** Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)

For this item of checklist, my review + https://github.com/openjournals/joss-reviews/issues/5347#issuecomment-1506886845 suffices for this to be marked as checked. 👍🏼
