### 2023-05-01 FZJ-PGI-12/QAOA.jl#27 @Abinashbunty
ArgumentError: Package ZygoteRules does not have ChainRulesCore in its dependencies
While executing `sherrington_kirkpatrick.ipynb` from [notebooks folder](https://github.com/FZJ-PGI-12/QAOA.jl/tree/master/notebooks), I get this error. Does the manifest file require some changes or is there some guide to resolve this issue?

Screeshot attached with more log information:

![image](https://user-images.githubusercontent.com/25522048/235527630-197605a5-4961-45bf-937b-d2fbe7b9e704.png)

Review tracking issue: openjournals/joss-reviews#5364
