### 2023-05-02 FZJ-PGI-12/QAOA.jl#27 @Abinashbunty
Maybe. I checked that based on [this](https://juliadiff.org/ChainRulesCore.jl/v1.0/converting_zygoterules.html), it might be useful to upgrade some existing _ZygotRules_ based functions to _ChainRules_.
