### 2023-05-05 FZJ-PGI-12/QAOA.jl#27 @timbode
> Maybe. I checked that based on [this](https://juliadiff.org/ChainRulesCore.jl/v1.0/converting_zygoterules.html), it might be useful to upgrade some existing _ZygotRules_ based functions to _ChainRules_.

The thing is, I am actually using `ChainRulesCore` and not `ZygoteRules`. So I'm not sure where this error comes from. Has this occured in other notebooks, too? What Julia version are you running?
