### 2023-05-22 FZJ-PGI-12/QAOA.jl#27 @Abinashbunty
@timbode It occurs on all the notebooks. I'm using Julia 1.8.5 from VS Code extension.
![image](https://github.com/FZJ-PGI-12/QAOA.jl/assets/25522048/2317faf4-854f-4b83-963d-c5be294af68a)
