### 2023-05-26 FZJ-PGI-12/QAOA.jl#28 @babreu-ncsa
Add community guidelines to documentation
*Guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support* is required in https://github.com/openjournals/joss-reviews/issues/5364
