### 2023-05-26 #5364 @babreu-ncsa
@timbode I created https://github.com/FZJ-PGI-12/QAOA.jl/issues/28 to address one of the **Documentation** check boxes. One comment regarding documentation is that I spent some time trying to understand why some methods are duplicated in the *Library* section until I realized that their input differs. If you could somehow make this more clear, it would be excellent.
