### 2023-05-26 FZJ-PGI-12/QAOA.jl#29 @babreu-ncsa
Add benchmark scripts for reproducibility
The **Benchmark** section of README.md shows a performance comparison between QAOA.jl and PennyLane. Although the data is in the assets folder, having the PennyLane script would be essential for direct reproducibility. Could you please add the script that was used for benchmarking? https://github.com/openjournals/joss-reviews/issues/5364
