### 2023-05-26 #5364 @babreu-ncsa
I've also created https://github.com/FZJ-PGI-12/QAOA.jl/issues/29 so users can reproduce the benchmarking claims. After these two are addressed, my review is done. Thanks!!
