+++ README.md
@@ -24,6 +24,11 @@ Our docs can be found [here](https://fzj-pgi-12.github.io/QAOA.jl/dev/). Example
 <img src="https://raw.githubusercontent.com/FZJ-PGI-12/QAOA.jl/master/assets/benchmarks.png" align="middle"/>
 
 
+## Support & Contributing
+
+In case you need support or have encountered a problem with the package, you are welcome to [create an issue on GitHub](https://github.com/FZJ-PGI-12/QAOA.jl/issues). If you would like to contribute to `QAOA.jl`, you can reach us via [PGI-12](https://www.fz-juelich.de/en/pgi/pgi-12).
+
+
 ## Citations
 
 If you are using code from this repository, please cite our work:
