### 2023-05-31 FZJ-PGI-12/QAOA.jl#28 @timbode
I've implemented the changes here: https://github.com/FZJ-PGI-12/QAOA.jl/commit/7a05f4144078221f9e6b0a23494b0cd893fdceab
