### 2023-05-31 FZJ-PGI-12/QAOA.jl#29 @timbode
@babreu-ncsa The notebook has been updated here: https://github.com/FZJ-PGI-12/QAOA.jl/blob/29-add-benchmark-scripts-for-reproducibility/notebooks/benchmark.ipynb

It should now be in a state where you just have to do `Run all` to reproduce the plot. Note, however, that the case with `p=5` and `n=20` will take something like 3 hours for PennyLane (on a M1 MacPro, that is). So you might want to change the `nrange` to something more economical.

Please let me know if you get results which are roughly in the same ball park. I got slightly different ones on my new machine with a different Julia version, which is to be expected though, I guess. If you give me green light, I'll do the merge.

Thanks!
