### 2023-06-02 FZJ-PGI-12/QAOA.jl#29 @timbode
I'll do the merge anyway now, should be fine.
