### 2023-06-02 f641f87
29 add benchmark scripts for reproducibility (#30)
* closes #29

* add missing docs file

---------

Co-authored-by: timbode <t.bode@Tims-MacBook-Pro-2.local>
