+++ docs/src/support.md
@@ -1,0 +1,3 @@
+# Support & Contributing
+
+In case you need support or have encountered a problem with the package, you are welcome to [create an issue on GitHub](https://github.com/FZJ-PGI-12/QAOA.jl/issues). If you would like to contribute to `QAOA.jl`, you can reach us via [PGI-12](https://www.fz-juelich.de/en/pgi/pgi-12).
