### 2023-06-27 #5364 @timbode
> * [ ]  Make a tagged release of your software, and list the version tag of the archived version here.

The version tag is [v1.0.1](https://github.com/FZJ-PGI-12/QAOA.jl/releases/tag/v1.0.1).
