+++ v1.0.1/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"examples/mean_field/#MFAOA","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"","category":"section"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"tip: Tip\nFor more details on the mean-field Approximate Optimization Algorithm, please consult our paper.","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"note: Note\nA Jupyter notebook related to this example is available in our examples folder. For a comparison between the QAOA and the mean-field AOA, have a look into our MaxCut example.","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"In close analogy to the QAOA, the mean-field Hamiltonian reads","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"beginalign\n    H(t) =  gamma(t)sum_i=1^N bigg h_i + sum_ji J_ij  n_j^z(t) bigg n_i^z(t) + beta(t) sum_i=1^N n_i^x(t)\nendalign","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"The mean-field evolution is then given by","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"beginalign\n    boldsymboln_i(p) = prod_k=1^p hat V_i^D(k) hat V_i^P(k) boldsymboln_i(0)\nendalign","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"where the initial spin vectors are boldsymboln_i(0) = (1 0 0)^T  forall i, and the rotation matrices hat V_i^DP are defined as","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"beginalign\nhat V_i^D(k) = \nbeginpmatrix\n1  0  0 \n0  cos(2 Delta_ibeta_k)  -sin(2 Delta_i beta_k) \n0  sin (2 Delta_i beta_k)  phantom-cos(2 Delta_i beta_k) \nendpmatrix\nendalign","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"and","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm","title":"Mean-Field Approximate Optimization Algorithm","text":"beginalign\nhat V_i^P(k) = \nbeginpmatrix\ncos(2m_i (t_k-1) gamma_k)  -sin(2m_i (t_k-1) gamma_k)  0 \nsin (2m_i (t_k-1) gamma_k)  phantom-cos(2m_i (t_k-1) gamma_k)  0 \n0  0  1\nendpmatrix\nendalign","category":"page"},{"location":"examples/mean_field/","page":"Mean-Field Approximate Optimization Algorithm"
[…]
