### 2023-06-27 #5364 @timbode
> @danielskatz I will try doing all of the above from a mobile device. Let’s see if it’s possible. If not, it will have to wait until the week after next.

Seems it all worked!
