### 2023-06-27 028ff98
minor changes in paper.md
