### 2023-06-27 #5364 @danielskatz
👋 @timbode - I'm suggesting some small changes in https://github.com/FZJ-PGI-12/QAOA.jl/pull/31

In addition, there's something wrong with the first reference. The DOI doesn't match the title, while the DOI that does match the title seems to be https://doi.org/10.5281/zenodo.2562111 but the authors don't match this archive, so I'm not sure what should be done to fix this.

Please go ahead and merge my PR (or let me know what you disagree with, and fix the first reference, then ping me here and I'll create a new proof.
